+++ Sources/EmbraceStorage/Records/Log/LogRecord.swift
@@ -30,6 +30,8 @@ public struct LogRecord {
 }
 
 extension LogRecord: FetchableRecord, PersistableRecord {
+    public static let databaseTableName: String = "logs"
+
     public static let databaseColumnDecodingStrategy = DatabaseColumnDecodingStrategy.convertFromSnakeCase
     public static let databaseColumnEncodingStrategy = DatabaseColumnEncodingStrategy.convertToSnakeCase
     public static let persistenceConflictPolicy = PersistenceConflictPolicy(insert: .replace, update: .replace)
@@ -78,20 +80,5 @@ extension LogRecord {
         static var body: Column { Column("body") }
         static var timestamp: Column { Column("timestamp") }
         static var attributes: Column { Column("attributes") }
-    }
-}
-
-extension LogRecord: TableRecord {
-    public static let databaseTableName: String = "logs"
-
-    internal static func defineTable(db: Database) throws {
-        try db.create(table: LogRecord.databaseTableName, options: .ifNotExists) { t in
-            t.primaryKey(Schema.identifier.name, .text).notNull()
-            t.column(Schema.processIdentifier.name, .integer).notNull()
-            t.column(Schema.severity.name, .integer).notNull()
-            t.column(Schema.body.name, .text).notNull()
-            t.column(Schema.timestamp.name, .datetime).notNull()
-            t.column(Schema.attributes.name, .text).notNull()
-        }
     }
 }